requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

openai==1.51.0
langchain==0.2.14
tiktoken==0.7.0

opensearch-py==2.4.2
boto3==1.34.0
requests-aws4auth==1.1.2

python-dotenv==1.0.1
pydantic==2.7.1
requests==2.32.3

# OCR 안 쓸 거면 paddleocr / paddlepaddle 전부 제거
# paddleocr==2.7.0
# paddlepaddle==2.6.2

gradio==4.44.0

loguru==0.7.2

pyjwt==2.8.0
passlib[bcrypt]==1.7.4

Pillow==10.3.0

pytest==8.2.2

httpx==0.27.2